Move some common elements into blocks file
