+++ packages/client-sdk/src/index.ts
@@ -15,8 +15,11 @@ export {
 } from "./iframe/position.js";
 
 export {
-  Events as IFrameEvents,
   IFRAME_NAME as MashIFrameName,
 } from "./iframe/IFrame.js";
+
+export {
+  Events as IFrameEvents,
+} from "./iframe/blocks.js";
 
 export { MashEvent } from "./events.js";
